Abstracted error handling and printing to a different format class
